pyproject.toml
[project]
name = "master-thesis-project"
version = "0.1.0"
description = "Dreamplan AI Financial Advisor"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "bs4>=0.0.2",
    "fastapi>=0.115.12",
    "haystack-ai>=2.13.1",
    "hf-xet>=1.1.2",
    "jupyter>=1.1.1",
    "langchain-openai>=0.3.16",
    "nltk>=3.9.1",
    "openai-agents>=0.0.4",
    "pandas>=2.2.3",
    "playwright>=1.52.0",
    "pre-commit>=4.2.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "radon>=6.0.1",
    "ragas>=0.2.15",
    "requests>=2.32.3",
    "ruff>=0.11.0",
    "sentence-transformers>=4.1.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
